% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis.R
\name{thesis_2021_pdf}
\alias{thesis_2021_pdf}
\title{Creates an R Markdown PDF Thesis document with the new style from 2021}
\usage{
thesis_2021_pdf(toc = TRUE, toc_depth = 3, highlight = "default", ...)
}
\arguments{
\item{toc}{A Boolean (TRUE or FALSE) specifying whether table of contents
should be created}

\item{toc_depth}{A positive integer}

\item{highlight}{Syntax highlighting style. Supported styles include
"default", "tango", "pygments", "kate", "monochrome", "espresso", "zenburn",
and "haddock". Pass NULL to prevent syntax highlighting.}

\item{...}{Additional parameters to pass to `pdf_book()`.}
}
\value{
A modified \code{pdf_document} based on the thesisdown template and
the huwiwidown template.
}
\description{
This is a function called in output in the YAML of the driver Rmd file to
specify using the Heidelberg University template.
}
\examples{
\dontrun{
 output: heididown::thesis_2021_pdf
}
}
